@@ -4,78 +4,73 @@ using Amazon.Lambda.Core;
 using Amazon.S3;
 using Amazon.S3.Model;
 using GrapeCity.Documents.Excel;
-using System;
-using System.Collections.Generic;
-using System.IO;
 using System.Net;
-using System.Threading.Tasks;
 
 // Assembly attribute to enable the Lambda function's JSON input to be converted into a .NET class.
 [assembly: LambdaSerializer(typeof(Amazon.Lambda.Serialization.SystemTextJson.DefaultLambdaJsonSerializer))]
 
-namespace ExcelExportAWSLambda2
+namespace ExcelExportAWSLambda2;
+
+public class Function
 {
-    public class Function
+    public async Task<APIGatewayProxyResponse> FunctionHandler(APIGatewayProxyRequest input, ILambdaContext
+ context)
     {
+        APIGatewayProxyResponse response;
 
-        public async Task<APIGatewayProxyResponse> FunctionHandler(APIGatewayProxyRequest input, ILambdaContext context)
+        try
         {
-            APIGatewayProxyResponse response;
+            // �N�G����������擾
+            string? queryString;
+            input.QueryStringParameters.TryGetValue("name", out queryString);
 
-            try
+            // ���[�N�V�[�g�ɒǉ�����e�L�X�g
+            string Message = string.IsNullOrEmpty(queryString)
+                ? "Hello, World!!"
+                : $"Hello, {queryString}!!";
+
+            //Workbook.SetLicenseKey("���i�ł܂��̓g���C�A���ł̃��C�Z���X�L�[��ݒ�");
+
+            Workbook workbook = new Workbook();
+            workbook.Worksheets[0].Range["A1"].Value = Message;
+
+            using (var ms = new MemoryStream())
             {
-                // �N�G����������擾
-                string queryString;
-                input.QueryStringParameters.TryGetValue("name", out queryString);
+                workbook.Save(ms, SaveFileFormat.Xlsx);
 
-                // ���[�N�V�[�g�ɒǉ�����e�L�X�g
-                string Message = string.IsNullOrEmpty(queryString)
-                    ? "Hello, World!!"
-                    : $"Hello, {queryString}!!";
-
-                //Workbook.SetLicenseKey("");
-
-                Workbook workbook = new Workbook();
-                workbook.Worksheets[0].Range["A1"].Value = Message;
-
-                using (var ms = new MemoryStream())
+                // S3�ɃA�b�v���[�h
+                AmazonS3Client client = new AmazonS3Client(RegionEndpoint.APNortheast1);
+                var request = new PutObjectRequest
                 {
-                    workbook.Save(ms, SaveFileFormat.Xlsx);
-
-                    // S3�ɃA�b�v���[�h
-                    AmazonS3Client client = new AmazonS3Client(RegionEndpoint.APNortheast1);
-                    var request = new PutObjectRequest
-                    {
-                        BucketName = "diodocs-export",
-                        Key = "Result.xlsx",
-                        InputStream = ms
-                    };
-
-                    await client.PutObjectAsync(request);
-                }
-
-                response = new APIGatewayProxyResponse
-                {
-                    StatusCode = (int)HttpStatusCode.OK,
-                    Body = "�t�@�C�����ۑ�����܂����B",
-                    Headers = new Dictionary<string, string> {
-                        { "Content-Type", "text/plain; charset=utf-8" }
-                    }
+                    BucketName = "diodocs-file-export",
+                    Key = "Result.xlsx",
+                    InputStream = ms
                 };
-            }
-            catch (Exception e)
-            {
-                response = new APIGatewayProxyResponse
-                {
-                    StatusCode = (int)HttpStatusCode.InternalServerError,
-                    Body = e.Message,
-                    Headers = new Dictionary<string, string> {
-                        { "Content-Type", "text/plain" }
-                    }
-                };
+
+                await client.PutObjectAsync(request);
             }
 
-            return response;
+            response = new APIGatewayProxyResponse
+            {
+                StatusCode = (int)HttpStatusCode.OK,
+                Body = "�t�@�C�����ۑ�����܂����B",
+                Headers = new Dictionary<string, string> {
+                { "Content-Type", "text/plain; charset=utf-8" }
+            }
+            };
         }
+        catch (Exception e)
+        {
+            response = new APIGatewayProxyResponse
+            {
+                StatusCode = (int)HttpStatusCode.InternalServerError,
+                Body = e.Message,
+                Headers = new Dictionary<string, string> {
+                { "Content-Type", "text/plain" }
+            }
+            };
+        }
+
+        return response;
     }
 }
